testcase-migration/requirements.txt
click
requests
python-dotenv
aiohttp
pyinstaller
